(refactor: split directory structure) 디렉터리 구조 나누기
