+++ src/main/java/com/weatherwhere/airservice/service/airforecast/GetAirForecastDataServiceImpl.java
@@ -8,11 +8,11 @@ import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.weatherwhere.airservice.domain.AirForecastEntity;
-import com.weatherwhere.airservice.domain.AirForecastId;
-import com.weatherwhere.airservice.dto.AirForecastDto;
-import com.weatherwhere.airservice.dto.SearchAirForecastDto;
-import com.weatherwhere.airservice.repository.AirForecastRepository;
+import com.weatherwhere.airservice.domain.airforecast.AirForecastEntity;
+import com.weatherwhere.airservice.domain.airforecast.AirForecastId;
+import com.weatherwhere.airservice.dto.airforecast.AirForecastDto;
+import com.weatherwhere.airservice.dto.airforecast.SearchAirForecastDto;
+import com.weatherwhere.airservice.repository.airforecast.AirForecastRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
